moved the BaseComponentUsingTaskQueueTestCase to the automated test package
